collection_app/requirements.txt
fire==0.4.0
nltk==3.7
pandas==1.3.2
pymorphy2==0.8
razdel==0.5.0
requests==2.22.0
psycopg2-binary>=2.8.4